fix: categoriAPI insert JwtCheck Token error

+++ src/main/java/io/devlog/blog/board/service/CategoryServiceImpl.java
@@ -42,7 +42,11 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public ResponseEntity<?> getCategories() {
         try {
-            List<Categories> categories = cateRepository.findByUserCateName(jwtService.checkJwt());
+            Long id = jwtService.checkJwt();
+            if (id == 0) {
+                return ResponseEntity.badRequest().body("Token error");
+            }
+            List<Categories> categories = cateRepository.findByUserCateName(id);
             List<CateDTO> cateDTOS = categories.stream()
                     .map(CateDTO::toDTO)
                     .collect(Collectors.toList());
@@ -57,7 +61,11 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public ResponseEntity<?> createCategory(CateDTO cateDTO) {
         try {
-            cateDTO.setUserUuid(jwtService.checkJwt());
+            Long id = jwtService.checkJwt();
+            if (id == 0) {
+                return ResponseEntity.badRequest().body("Token error");
+            }
+            cateDTO.setUserUuid(id);
             cateRepository.save(cateDTO.toEntity());
             Optional<Categories> category = cateRepository.findByCateName(cateDTO.getCateName());
             if (category.isEmpty()) {
@@ -73,6 +81,10 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public ResponseEntity<?> updateCateName(CateDTO cateDTO) {
         try {
+            Long id = jwtService.checkJwt();
+            if (id == 0) {
+                return ResponseEntity.badRequest().body("Token error");
+            }
             Optional<Categories> existingCategory = cateRepository.findByCateUuid(cateDTO.getCateUuid());
             if (existingCategory.isPresent()) {
                 Categories category = existingCategory.get();
@@ -92,6 +104,10 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public ResponseEntity<?> updateCategory(List<CateDTO> cateDTOS) {
         try {
+            Long id = jwtService.checkJwt();
+            if (id == 0) {
+                return ResponseEntity.badRequest().body("Token error");
+            }
             for (CateDTO cateDTO : cateDTOS) {
                 Optional<Categories> existingCategory = cateRepository.findByCateUuid(cateDTO.getCateUuid());
                 if (!existingCategory.isEmpty()) {
@@ -114,6 +130,10 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public ResponseEntity<?> deleteCategory(Long cateUuid) {
         try {
+            Long id = jwtService.checkJwt();
+            if (id == 0) {
+                return ResponseEntity.badRequest().body("Token error");
+            }
             if (cateRepository.deleteByCategoriId(cateUuid) == 1) {
                 return ResponseEntity.status(200).body("category deleted");
             } else {
